refactor: 아이콘 컴포넌트 atoms 폴더로 이동
