+++ client/src/components/Modal/ShareFormModal/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ToggleButton from "components/ToggleButton";
-import Icon from "components/Icon";
+import Icon from "components/atoms/Icon";
 import Button from "components/atoms/Button";
 import theme from "styles/theme";
 import * as S from "./style";
