+++ client/src/components/atoms/Icon/style.ts
